requirements.txt
adal==1.2.7
adlfs==2022.4.0
aiobotocore==2.2.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.23.1
azure-datalake-store==0.0.52
azure-identity==1.9.0
azure-storage-blob==12.11.0
boto3==1.21.21
botocore==1.24.21
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
frozenlist==1.3.0
fsspec==2022.3.0
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
jmespath==1.0.0
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
multidict==6.0.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
portalocker==2.4.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
s3fs==2022.3.0
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
